Experimental support for using SpaceBars templates in content pages
WIP. Does not – among other things, probably – handle and editor
referring to the surrounding template name in content, which will cause
a stack overflow.

We will probably not use this feature, but opt for a simpler solution for
including components (this experiment came about when implementing the
donate button on the donate content page).